requirements.txt
# Dependências Python - SenaChat Backend

# Framework web
Flask==3.0.0
Flask-CORS==4.0.0

# Google Gemini AI
google-genai==1.0.0

# Variáveis de ambiente
python-dotenv==1.0.0

# Servidor WSGI para produção (opcional)
# gunicorn==21.2.0
